pip_requirements.txt
networkx
seaborn
matplotlib
pandas
scikit-learn
numpy
scipy
google-cloud-bigquery
google-cloud-storage
dbt-fal[bigquery]
dbt-bigquery
notebook
torch
torchvision
torchaudio
torch_geometric
-f https://data.pyg.org/whl/torch-2.2.1+cu121.html
pyg_lib
torch_scatter
torch_sparse
torch_cluster
torch_spline_conv
torcheval